HU-01-login (#51)
* feat: Apply AppTrueQTheme to LoginScreen

The `LoginScreen` composable is now wrapped with the `AppTrueQTheme` to ensure it uses the custom application theme for a consistent visual style.

* feat: Make login and register buttons full-width

The "Iniciar Sesión" and "Regístrate" buttons on the `LoginScreen` have been updated to span the full width of the screen. This is achieved by adding the `modifier = Modifier.fillMaxWidth()` to both `Button` components.

---------

Co-authored-by: Wilber Ticona <wilber.ticona@outlook.com>